Add a helper function to preload state with an initial state that works with optismitic
When redux initializes it passes a couple of actions to each reducer.
This initialization was sets the "isReady" flag. If we have an
initial state coming from outside of the reducer, however, then on the
actual @@INIT pass the isReady flag would be true and the state be
defined even though it isn't actually initialized.

In this instance, use the preloadState function to pass in the initial
value so it's already set up for optimistic.

Fixes #16