Add documentation on transaction behavior
This was pretty confusing for me as a first-time user and I had to read the source code and do several experiments to figure this out. I'm submitting this change to the README that adds a section on the specific behavior of `BEGIN`, `COMMIT`, and `REVERT` and also includes a recommendation on how to use them effectively.

I also cleaned up the formatting. FYI GIthub doesn't recognize headings unless there's a space following the `###`