fix issue with commit/revert of invalid transaction id
A bug exists where the check for an invalid transaction id comes too
late, which means that the check actually does find an action with that
transaciton id - the invalid one that was just pushed.

The fix is to check whether the transaction id is valid earlier, and
throw an Error (rather than just logging it).

Fixes #32