change method for detemrining whether state is already wrapped in reducer
Prevously we would only wrap the state in the redux-optimistic-ui wrapper if the passed
in state was undefined, or this was the `@@redux/INIT` action. However when initial state is
passed to `createStore` other middleware, notably the Redux devtools instrumentation, will
trigger an `@@INIT` action with the unwrapped state. This wouldn't match our check, and so
would later throw an error as it wasn't wrapped state.

The check for whether wrapped state has been passed to our reducer has been rewritten, instead
we now verify that the properties in the state match what our wrapped state looks like.